+++ NEWS.md
@@ -1,4 +1,4 @@
-# snvecR (development version)
+# snvecR 3.10.0
 * Added support for reading ZB23.Rxx and pre-computed PT-ZB18a solutions.
 * snvec() now also returns lpx by default.
 * CRAN tests no longer rely on downloading the solutions.